backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
RPi.GPIO==0.7.1
mfrc522==0.0.7
spidev==3.6
supabase==2.3.4
python-dotenv==1.0.0


